proximator schema - enable index
