+++ python/src/gpx2db/sql/proximity-calc/0200_create_segments_table.sql
@@ -1,5 +1,5 @@
-drop table if exists newsegments;
-create table newsegments (
+drop table if exists {schema}.newsegments;
+create table {schema}.newsegments (
 	segment_id integer primary key,
 	track_id integer not null,
 	numpoints integer not null,
