+++ python/src/gpx2db/sql/proximity-calc/0300_insert_segments.sql
@@ -1,9 +1,9 @@
-insert into newsegments
+insert into {schema}.newsegments
 select
 	segment_id,
 	track_id,
 	count(point_id),
 	ST_MakeLine(wkb_geometry order by point_id)
-from newpoints
+from {schema}.newpoints
 where track_id = {}
 group by track_id, segment_id;
