+++ python/src/gpx2db/sql/proximity-calc/0400_create_segments_table_idx.sql
@@ -1,3 +1,3 @@
 CREATE INDEX newsegments_geom_idx
-    ON newsegments USING gist
+    ON {schema}.newsegments USING gist
     (wkb_geometry);
