+++ python/src/gpx2db/sql/proximity-calc/1200_create_intersect_table_idx.sql
@@ -2,6 +2,6 @@ drop index if exists intersect_table_idx_cid;
 drop index if exists intersect_table_idx_segid;
 drop index if exists intersect_table_idx_geom;
 
-create index intersect_table_idx_cid on circle_segment_intersections (circle_id);
-create index intersect_table_idx_segid on circle_segment_intersections (segment_id);
-CREATE INDEX intersect_table_idx_geom ON circle_segment_intersections USING gist(wkb_geometry);
+create index intersect_table_idx_cid on {schema}.circle_segment_intersections (circle_id);
+create index intersect_table_idx_segid on {schema}.circle_segment_intersections (segment_id);
+CREATE INDEX intersect_table_idx_geom ON {schema}.circle_segment_intersections USING gist(wkb_geometry);
