+++ python/src/gpx2db/sql/proximity-calc/1310_insert_circles.sql
@@ -1,7 +1,7 @@
-insert into circles (
+insert into {schema}.circles (
     select
         tp.point_id,
         ST_Buffer(tp.wkb_geometry::geography, {})::geometry
-    from newpoints as tp
+    from {schema}.newpoints as tp
     where tp.track_id = {}
 );
