+++ python/src/gpx2db/sql/proximity-calc/3300_view_count_linestrings.sql
@@ -1,10 +1,10 @@
-drop view if exists count_ls;
-create view count_ls as
+drop view if exists {schema}.count_ls;
+create view {schema}.count_ls as
 select
 	circle_id,
 	count(*) as num
 from
-	intersections_gtype
+	{schema}.intersections_gtype
 where
 	geom_type = 'ST_LineString'
 group by
